refactor: rename and replace chat media page view component for improved clarity
